+++ Revit_Core_Engine/Query/AdjustedLocationCurve.cs
@@ -1,0 +1,84 @@
+﻿/*
+ * This file is part of the Buildings and Habitats object Model (BHoM)
+ * Copyright (c) 2015 - 2020, the respective contributors. All rights reserved.
+ *
+ * Each contributor holds copyright over their respective contributions.
+ * The project versioning (Git) records all such contribution source information.
+ *                                           
+ *                                                                              
+ * The BHoM is free software: you can redistribute it and/or modify         
+ * it under the terms of the GNU Lesser General Public License as published by  
+ * the Free Software Foundation, either version 3.0 of the License, or          
+ * (at your option) any later version.                                          
+ *                                                                              
+ * The BHoM is distributed in the hope that it will be useful,              
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of               
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                 
+ * GNU Lesser General Public License for more details.                          
+ *                                                                            
+ * You should have received a copy of the GNU Lesser General Public License     
+ * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
+ */
+
+using Autodesk.Revit.DB;
+using BH.Engine.Adapters.Revit;
+using BH.Engine.Geometry;
+using BH.oM.Adapters.Revit.Settings;
+using BH.oM.Geometry;
+using BH.oM.Physical.Elements;
+using BH.oM.Reflection;
+
+namespace BH.Revit.Engine.Core
+{
+    public static partial class Query
+    {
+        /***************************************************/
+        /****              Public methods               ****/
+        /***************************************************/
+        
+        public static ICurve AdjustedLocationCurveFraming(this IFramingElement framingElement, FamilyInstance familyInstance, RevitSettings settings = null)
+        {
+            settings = settings.DefaultIfNull();
+
+            ICurve curve = framingElement?.Location;
+            if (curve == null || curve.ILength() <= settings.DistanceTolerance)
+            {
+                framingElement.FramingCurveNotFoundWarning();
+                return null;
+            }
+
+            BH.oM.Geometry.Line line = curve as BH.oM.Geometry.Line;
+            if (line == null)
+            {
+                familyInstance.NonLinearFramingOffsetWarning();
+                return curve;
+            }
+
+            Output<Vector, Vector> offsets = familyInstance.FramingOffsetVectors();
+            Vector startOffset = -offsets.Item1;
+            Vector endOffset = -offsets.Item2;
+
+            double startOffsetLength = startOffset.Length();
+            double endOffsetLength = endOffset.Length();
+            if (startOffsetLength > Tolerance.Distance || endOffsetLength > Tolerance.Distance)
+            {
+                if ((startOffset - endOffset).Length() > settings.DistanceTolerance)
+                {
+                    //throw warning and reset offsets
+                }
+
+                Transform transform = familyInstance.GetTotalTransform();
+                Vector yOffsetStart = new Vector { X = transform.BasisY.X * startOffset.Y, Y = transform.BasisY.Y * startOffset.Y, Z = transform.BasisY.Z * startOffset.Y };
+                Vector zOffsetStart = new Vector { X = transform.BasisZ.X * startOffset.Z, Y = transform.BasisZ.Y * startOffset.Z, Z = transform.BasisZ.Z * startOffset.Z };
+                Vector yOffsetEnd = new Vector { X = transform.BasisY.X * endOffset.Y, Y = transform.BasisY.Y * endOffset.Y, Z = transform.BasisY.Z * endOffset.Y };
+                Vector zOffsetEnd = new Vector { X = transform.BasisZ.X * endOffset.Z, Y = transform.BasisZ.Y * endOffset.Z, Z = transform.BasisZ.Z * endOffset.Z };
+                curve = new BH.oM.Geometry.Line { Start = line.Start.Translate(yOffsetStart + zOffsetStart), End = line.End.Translate(yOffsetEnd + zOffsetEnd) };
+            }
+
+            return curve;
+        }
+
+        /***************************************************/
+    }
+}
+
